# app/swagger.yaml
swagger: "2.0"
info:
  title: Medical Records API
  description: API untuk mengelola data rekam medis pasien, pengguna, laporan, dan dokumen medis lainnya.
  version: 1.0.0
host: localhost:5000
basePath: /
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Masukkan token JWT dengan format: 'Bearer {token}'"
security:
  - bearerAuth: []
...
definitions:
...
  Assesment:
    type: object
    properties:
      id:
        type: integer
      tanggal:
        type: string
        format: date-time
      perawat:
        type: string
      data:
        type: object
        description: "Struktur JSON dari hasil asesmen."
...
  Query:
    type: object
    properties:
      query:
        type: string
        description: "Teks atau kueri untuk diproses oleh AI atau pencarian."
paths:
...
  /assesments/:
    get:
      tags:
        - Assesments
      summary: Mendapatkan semua data asesmen
      security: []
      responses:
        '200':
          description: Sukses
    post:
      tags:
        - Assesments
      summary: Membuat asesmen baru dengan AI
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              query:
                type: string
              perawat:
                type: string
      responses:
        '201':
          description: Asesmen berhasil dibuat
  /assesments/search:
    post:
      tags:
        - Assesments
      summary: Mencari data asesmen
      security: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Query'
      responses:
        '200':
          description: Hasil pencarian
  /assesments/{assesment_id}:
    parameters:
      - name: assesment_id
        in: path
        required: true
        type: integer
    get:
      tags:
        - Assesments
      summary: Mendapatkan detail asesmen
      security: []
      responses:
        '200':
          description: Sukses
    put:
      tags:
        - Assesments
      summary: Memperbarui asesmen
      security: []
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Assesment'
      responses:
        '200':
          description: Asesmen berhasil diperbarui
    delete:
      tags:
        - Assesments
      summary: Menghapus asesmen
      security: []
      responses:
        '200':
          description: Asesmen berhasil dihapus
...